Updated Raspy AI
- Updated Raspy AI component for a drawer instead of a dialog
- Added hoc pattern for multiple implementations

Added support for reporting

- Added support to ensure users can view reports on their properties.
- Lifted data out of Raspy AI to limit raspy for search and only to generate
recommended actions from the reports.
- Also we would like to use raspy to focus more on property management but still
co-ordinating on this aspect.

Remove Raspy AI features

- After a lot of consideration, we decided to put a pause on the Raspy AI feature. Instead of providing recommendations,
we want the Raspy AI feature to be a complete overhaul. Hence, removing raspy ai calls from the properties page or the
reporting and analytics page.

- Refactor the charts into reporting folder structure so that the components better chunk themselves during build

- We want the raspy ai support to be more geared towards some real world problems, like providing support for property owners
to find tenants of their liking and support booking for maintenance and other tools. We would also like the AI to support
the ability to provide some regular household maintenance tips when needed, and all of that is on a different direction that what we
are using Raspy AI for here. Hence, removed all code atm to support this ask.
