+++ src/common/RowHeader.jsx
@@ -4,20 +4,6 @@ import dayjs from "dayjs";
 
 import { Box, Stack, Typography } from "@mui/material";
 
-/**
- * RowHeader
- *
- * used to build out header components with custom css decorators. Some css decorators bleed into title variants.
- * Eg, fontWeight is overridden in a few places with this feature.
- *
- * @param {string} title - The header title
- * @param {string} caption - The caption for the title
- * @param {boolean} showDate - The true or false value to display date
- * @param {dayjs} createdDate - The date value of the created timestamp
- * @param {Object} sxProps - Additional wrapper styles based on parent components.
- * @param {Object} children - Can include jsx components to add more data / details to the RowHeader Component.
- *
- */
 export default function RowHeader({
   title,
   caption,
