+++ src/common/ScrollTop/ScrollTopProvider.jsx
@@ -4,15 +4,9 @@ import { KeyboardArrowUp } from "@mui/icons-material";
 import { Box, Fab } from "@mui/material";
 import ScrollTop from "common/ScrollTop/ScrollToTop";
 
-/**
- * ScrollTopProvider Component
- *
- * Component built with the provider pattern to create a FAB button
- * that allows the user to navigate to the top when clicked. The div
- * id of "back-to-top-anchor" is the id that allows the fab to scroll
- * to the top of the page.
- *
- */
+// ScrollTopProvider ...
+// defines a provider function that creates a FAB button to allow
+// users to scroll to the top when selected.
 export default function ScrollTopProvider({ children }) {
   return (
     <>
