+++ src/common/utils.jsx
@@ -15,6 +15,8 @@ export const PropertiesRoutePath = "properties";
 export const PropertiesRouteUri = "/rent/properties";
 export const RentalRoutePath = "rental";
 export const RentalRouteUri = "/rent/rental";
+export const PropertiesReportingPath = "reporting";
+export const PropertiesReportingRouteUri = "/rent/reporting";
 export const SettingsRoutePath = "settings";
 export const SettingsRouteUri = "/rent/settings";
 export const PropertyRoutePath = "property/:id";
