+++ src/features/Rent/components/ProfileDetails/ProfileDetails.jsx
@@ -42,7 +42,7 @@ export default function ProfileDetails() {
   const {
     data: latestSubscription = {},
     isLoading: isSubscriptionDetailsLoading,
-  } = useGetLatestSubscriptionByEmailQuery(user.email, {
+  } = useGetLatestSubscriptionByEmailQuery(user?.email, {
     skip: !user?.email,
   });
